Allow Azure Resource Group template to be contained in package
