+++ source/Calamari/Deployment/Conventions/ExtractPackageToStagingDirectoryConvention.cs
@@ -1,4 +1,5 @@
-﻿using Calamari.Integration.FileSystem;
+﻿using System;
+using Calamari.Integration.FileSystem;
 using Calamari.Integration.Packages;
 
 namespace Calamari.Deployment.Conventions
@@ -12,13 +13,7 @@ namespace Calamari.Deployment.Conventions
 
         protected override string GetTargetPath(RunningDeployment deployment, PackageMetadata metadata)
         {
-            var packageExtractionPathVariable = deployment.Variables[SpecialVariables.Action.Azure.PackageExtractionPath];
-
-            // The PackageExtractionPath variable will always be provided by the OD server, but just in case Calamari is run
-            // stand-alone, we will fall-back to a temporary path
-            return !string.IsNullOrWhiteSpace(packageExtractionPathVariable)
-                ? packageExtractionPathVariable
-                : fileSystem.CreateTemporaryDirectory();
+            return Environment.CurrentDirectory;
         }
     }
 }
